requirements.txt
langchain
langchain-core
langchain-community
langchain-ollama
python-dotenv
requests
numpy
transformers
torch
numpy
